Refactor CommandBuffer

+++ Source/CommandBuffer.h
@@ -51,7 +51,7 @@ public:
 	virtual void setClearDepth(float depth) = 0;
 	virtual void setClearStencil() = 0;
 
-	virtual void draw(uint32_t first, uint32_t num) = 0;
+	virtual void draw(size_t first, size_t num) = 0;
 
 private:
 	std::shared_ptr<Program> program;
