+++ Source/Vulkan/VKCommandBuffer.h
@@ -21,5 +21,5 @@ public:
 	void setClearDepth(float depth) override;
 	void setClearStencil() override;
 
-	void draw(uint32_t first, uint32_t num) override;
+	void draw(size_t first, size_t num) override;
 };
